+++ src/main/java/com/project/contap/service/MypageService.java
@@ -1,6 +1,5 @@
 package com.project.contap.service;
 
-import com.google.common.collect.Sets;
 import com.project.contap.dto.*;
 import com.project.contap.exception.ContapException;
 import com.project.contap.exception.ErrorCode;
@@ -30,17 +29,18 @@ public class MypageService {
     private final CardRepository cardRepository;
     private final HashTagRepositoty hashTagRepositoty;
     private final ImageService imageService;
+    private final UserService userService;
 
     // 회원 정보 가져오기
     // 가져오는 값 : 기본 회원정보(앞면카드), 모든 뒷면카드
     public UserInfoDto getMyInfo(User requestUser) {
 
-        User user = checkUserAuthority(requestUser);
+        User user = userService.checkUserAuthority(requestUser);
         List<Card> userCards = user.getCards();
         List<BackResponseCardDto> cardDtoList = new ArrayList<>();
 
         for(Card card: userCards) {
-            BackResponseCardDto cardDto = makeBackResponseCardDto(card,user);
+            BackResponseCardDto cardDto = makeBackResponseCardDto(card, user);
             cardDtoList.add(cardDto);
         }
 
@@ -49,6 +49,7 @@ public class MypageService {
                 .password(user.getPw())
                 .userName(user.getUserName())
                 .profile(user.getProfile())
+                .field(user.getField())
                 .authorityEnum(user.getAuthorityEnum())
                 .cardDtoList(cardDtoList).build();
     }
@@ -56,7 +57,7 @@ public class MypageService {
     //앞카드 수정
     @Transactional
     public FrontResponseCardDto modifyFrontCard(FrontRequestCardDto frontRequestCardDto, User requestUser) throws IOException {
-        User user = checkUserAuthority(requestUser);
+        User user = userService.checkUserAuthority(requestUser);
 
         // nickname 변경했을 경우 중복체크
         if(!user.getUserName().equals(frontRequestCardDto.getUserName())) {
@@ -77,9 +78,11 @@ public class MypageService {
         }
 
         String requestTagStr = frontRequestCardDto.getHashTagsStr();
-        Set<String> sets = null;
+        Set<String> sets = new HashSet<>();
         if(requestTagStr.contains(SPLIT_CHAR)) {
-            sets = new HashSet<>(Sets.newHashSet(requestTagStr.split(SPLIT_CHAR)));
+            Collections.addAll(sets, requestTagStr.split(SPLIT_CHAR));
+        }else if(requestTagStr.length()>0) {
+            sets.add(requestTagStr);
         }
 
         List<HashTag> hashTagList = hashTagRepositoty.findAllByNameIn(sets);
@@ -104,7 +107,7 @@ public class MypageService {
     @Transactional
     public BackResponseCardDto createBackCard(BackRequestCardDto backRequestCardDto, User requestUser) {
 
-        User user = checkUserAuthority(requestUser);
+        User user = userService.checkUserAuthority(requestUser);
 
         int cardSize = user.getCards().size();
         if( cardSize >= 10 )
@@ -120,7 +123,6 @@ public class MypageService {
         card.setLink(backRequestCardDto.getLink());
         card = cardRepository.save(card);
 
-        //다른 user 뒷면 볼 수 있는 권한
         user.setAuthorityEnum(AuthorityEnum.CAN_OTHER_READ);
 
         //response
@@ -131,7 +133,7 @@ public class MypageService {
     @Transactional
     public BackResponseCardDto modifyBackCard(Long cardId, BackRequestCardDto backRequestCardDto, User requestUser) {
 
-        User user = checkUserAuthority(requestUser);
+        User user = userService.checkUserAuthority(requestUser);
         Card card = cardRepository.findById(cardId).orElse(null);
         if(card == null)
             throw new ContapException(ErrorCode.NOT_FOUND_CARD); //해당 카드를 찾을 수 없습니다.
@@ -148,7 +150,7 @@ public class MypageService {
     @Transactional
     public BackResponseCardDto deleteBackCard(Long cardId, User requestUser) {
 
-        User user = checkUserAuthority(requestUser);
+        User user = userService.checkUserAuthority(requestUser);
         Card card = cardRepository.findById(cardId).orElse(null);
         if(card == null)
             throw new ContapException(ErrorCode.NOT_FOUND_CARD); //해당 카드를 찾을 수 없습니다.
@@ -178,14 +180,14 @@ public class MypageService {
     }
 
     // 사용자 권한 체크
-    public User checkUserAuthority(User requestUser) {
-        if(requestUser == null)
-            throw new ContapException(ErrorCode.USER_NOT_FOUND); //회원 정보를 찾을 수 없습니다.
-
-        User user = userRepository.findById(requestUser.getId()).orElse(null);
-        if (user.getEmail()!=null && !user.isWritedBy(requestUser))
-            throw new ContapException(ErrorCode.ACCESS_DENIED); //권한이 없습니다.
-
-        return user;
-    }
+//    public User checkUserAuthority(User requestUser) {
+//        if(requestUser == null)
+//            throw new ContapException(ErrorCode.USER_NOT_FOUND); //회원 정보를 찾을 수 없습니다.
+//
+//        User user = userRepository.findById(requestUser.getId()).orElse(null);
+//        if (user.getEmail()!=null && !user.isWritedBy(requestUser))
+//            throw new ContapException(ErrorCode.ACCESS_DENIED); //권한이 없습니다.
+//
+//        return user;
+//    }
 }
